refactor(racingcar.*Test): 테스트 패키지를 프로덕션 패키지와 동일하게 수정
test class와 src/test, src/main 부터 상대 경로가 같은(하위 패키지 구조가 같은) class는 따로 import 없이 사용할 수 있다
